+++ src/pages/index.page.tsx
@@ -1,54 +1,37 @@
 import type { NextPage } from "next";
-import { SWRConfig } from "swr";
-import { Fallback } from "model";
-import { getStaticProps } from "./index.hooks";
-import { useCallback, useEffect, useState } from "react";
-import { useHandleDisplay } from "hooks/useHandleDisplay";
-import { Layout } from "Layout";
-import { TitleListContainer } from "components/TitleListContainer";
-import { ModalForSearch } from "components/Modal";
-import { SearchScreen } from "components/Modal/SearchScreen";
-import { Map } from "components/Map";
-import { GoogleMapComponent } from "components/Map/GoogleMapComponent";
+import Link from "next/link";
+import Image from "next/image";
+import { Box } from "@mui/system";
 
-export { getStaticProps };
-
-export const App: NextPage<{ fallback: Fallback }> = (props) => {
-  const { fallback } = props;
-  const [titleList, setTitleList] = useState<string[]>([]);
-  const [isShow, setIsShow] = useState(false);
-  const handleDisplay = useHandleDisplay(setIsShow);
-
-  const handleKeyDown = useCallback(
-    (ev: KeyboardEvent) => {
-      if (ev.key === "k" && (ev.ctrlKey || ev.metaKey)) {
-        handleDisplay();
-      }
-    },
-    [handleDisplay]
-  );
-
-  useEffect(() => {
-    document.addEventListener("keydown", handleKeyDown);
-  }, [handleKeyDown]);
-
+export const App: NextPage = (props) => {
   return (
-    <SWRConfig value={{ fallback }}>
-      <Layout>
-        <TitleListContainer
-          isShow={isShow}
-          setTitleList={setTitleList}
-          handleDisplay={handleDisplay}
-          titleList={titleList}
-        >
-          <ModalForSearch setIsShow={setIsShow} handleDisplay={handleDisplay}>
-            <SearchScreen setTitleList={setTitleList} setIsShow={setIsShow}/>
-          </ModalForSearch>
-        </TitleListContainer>
-
-        <Map GoogleMapComponent={<GoogleMapComponent titleList={titleList} />} />
-      </Layout>
-    </SWRConfig>
+    <Box
+      sx={{
+        maxHeight: "100vh",
+        px: 38,
+        py: 15,
+      }}
+    >
+      <Box sx={{ height: "60vh", display: "flex" }}>
+        <Link href="/map" passHref>
+          <Box
+            sx={{
+              position: "relative",
+              width: "100%",
+              height: "97%",
+              ":hover": { cursor: "pointer", boxShadow: 5 },
+            }}
+          >
+            <Image src="/Aincrad.png" layout="fill" alt="Aincrad" objectFit="contain" />
+          </Box>
+        </Link>
+      </Box>
+      <Box height="10vh">
+        <section>
+          <h1>日本地図は、アニメ聖地と共に進化する。さぁ、完成させよAniMap</h1>
+        </section>
+      </Box>
+    </Box>
   );
 };
